Changes to get the stemming code to compile.  Back-fit the MacOS changes to Windows.  Should be one code base now.
git-svn-id: https://groke.otago.ac.nz/ant/ant@74 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/search_engine_btree_leaf.h
@@ -8,7 +8,7 @@
 class ANT_search_engine_btree_leaf
 {
 public:
-	long collection_frequency;
+	long long collection_frequency;
 	long document_frequency;
 	long long postings_position_on_disk;
 	long docid_length;
